requirements.txt
mysql-connector-python
phpserialize
